refactor(app): give App.tsx's module furniture and OAuth session real homes
Structure review, finding 16 — the low-risk half, which the review calls out
as independent of everything else.

App.tsx opened with ~110 lines that were not App: two starter configs, the
results half's `lazy()` boundary and its wrapper, the run-chunk preloader, and
the page-level banner row. Moved to `data/starter-configs.ts`,
`app/ResultsPane.tsx`, `app/preload-run-chunks.ts` and `app/AppBanners.tsx`, so
the file now starts at the component. 1,887 -> 1,753 lines.

`preloadRunChunks` is its own module rather than sitting beside the `lazy()` it
warms — `react/only-export-components` refuses a file exporting both a
component and a function, the same reason the repo's `*-hooks.ts` files exist.
Both modules now carry the note that their two dynamic-import specifiers must
stay identical, since a preloader warming a different string than the boundary
loads is a bug whose only symptom is a slow first Run.

`useOAuthSession` takes the sign-in session: `signedIn`, `authUser`, the
`authState` derivation that sat 1,000 lines from its inputs, `onSignOut`, and
`oauthConfigured` in place of six separate `Boolean(OAUTH_CONFIG)` spellings.

`onSignIn` deliberately did NOT move, and the hook's header records why: it has
to decide what the user returns TO, which reads App's `result` and the share
hook's return-hash builder — and the share hook takes this hook's setters,
since it is what handles the OAuth callback. That is a real ordering cycle, not
a cosmetic one; oxlint's `react/immutability` rejected the version that read
those bindings early, which is the rule earning its keep. So the session state
lives in the hook and the sign-in policy stays where its inputs are.

Also drops three exported values nobody consumed — `keysLive`, `focusEditor`
and `focusResults` on the 573-line `useKeyboardLandings`. Verified before
removing: outside the hook those names appear only in other files' PROSE. App
destructured 9 of the 12; the return is 9 now. `keysLive`'s doc claimed "App
passes it on to any binding of its own", and App has had no such binding for
some time.

The three remaining state clusters (`useConfigDocument`, `useRepoProvenance`,
`useRunViewSelection`) are untouched — the review asks for them one at a time,
and they are not warm-ups.

968 tests green, lint and typecheck clean.
